requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.3.1
billiard==3.3.0.23
celery==3.1.25
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
Django==1.8.4
django-celery==3.1.17
django-haystack==3.0
django-redis==4.8.0
django-tinymce==2.6.0
fdfs-client-py @ file:///home/python/fdfs_client-py-master.zip
idna==2.10
itsdangerous==1.1.0
jieba==0.42.1
kombu==3.0.37
mutagen==1.45.1
mysqlclient==2.0.3
Pillow==8.1.0
pycparser==2.20
pycryptodomex==3.9.4
pyOpenSSL==19.1.0
python-alipay-sdk==1.8.0
pytz==2021.1
redis==2.10.6
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
uWSGI==2.0.19.1
Whoosh==2.7.4
